Changed stress type from `int` to `long` (this is what most of these changes are about).
Changed solution type to `long`.

Updated `Solver` to solve Part 2.

Added Part 2 solution to program output.

Updated `KeepAwayGame` to use a user-provided stress management strategy.

+++ 2022/day 11/Solution/ItemTest.cs
@@ -13,7 +13,7 @@ public class ItemTest
         FalseTarget = falseTarget;
     }
 
-    public int Perform(int itemWorryLevel) =>
+    public int Perform(long itemWorryLevel) =>
         itemWorryLevel % Divisor == 0 
             ? TrueTarget 
             : FalseTarget;
