+++ 2022/day 11/Solution/Monkey.cs
@@ -2,7 +2,7 @@
 
 public class Monkey
 {
-    public List<int> Items { get; init; } = new();
+    public List<long> Items { get; init; } = new();
 
     public Operation Operation { get; init; } = new("0 + 0");
     public ItemTest ItemTest { get; init; } = new(0, 0, 0);
