+++ 2022/day 11/Tests/KeepAwayGameTests.cs
@@ -6,29 +6,29 @@ public class KeepAwayGameTests
     public void DoRound_CausesMonkeysToExchangeItems_GivenAdventOfCodeExample()
     {
         // Arrange
-        var defaultStressManagementStrategy = (int itemWorryLevel) => itemWorryLevel / 3;
+        var defaultStressManagementStrategy = (long itemWorryLevel) => itemWorryLevel / 3;
         var game = new KeepAwayGame(defaultStressManagementStrategy);
         game.Monkeys.Add(new()
         {
-            Items = new List<int> { 79, 98 },
+            Items = new List<long> { 79, 98 },
             Operation = new Operation("old * 19"),
             ItemTest = new ItemTest(23, 2, 3),
         });
         game.Monkeys.Add(new()
         {
-            Items = new List<int> { 54, 65, 75, 74 },
+            Items = new List<long> { 54, 65, 75, 74 },
             Operation = new Operation("old + 6"),
             ItemTest = new ItemTest(19, 2, 0),
         });
         game.Monkeys.Add(new()
         {
-            Items = new List<int> { 79, 60, 97 },
+            Items = new List<long> { 79, 60, 97 },
             Operation = new Operation("old * old"),
             ItemTest = new ItemTest(13, 1, 3),
         });
         game.Monkeys.Add(new()
         {
-            Items = new List<int> { 74 },
+            Items = new List<long> { 74 },
             Operation = new Operation("old + 3"),
             ItemTest = new ItemTest(17, 0, 1),
         });
@@ -37,8 +37,8 @@ public class KeepAwayGameTests
         game.DoRound();
 
         // Assert
-        game.Monkeys[0].Items.Should().BeEquivalentTo(new List<int> { 20, 23, 27, 26 });
-        game.Monkeys[1].Items.Should().BeEquivalentTo(new List<int> { 2080, 25, 167, 207, 401, 1046 });
+        game.Monkeys[0].Items.Should().BeEquivalentTo(new List<long> { 20, 23, 27, 26 });
+        game.Monkeys[1].Items.Should().BeEquivalentTo(new List<long> { 2080, 25, 167, 207, 401, 1046 });
         game.Monkeys[2].Items.Should().BeEmpty();
         game.Monkeys[3].Items.Should().BeEmpty();
     }
@@ -47,29 +47,29 @@ public class KeepAwayGameTests
     public void DoRound_TracksHowManyTimesMonkeysInspectItems_GivenAdventOfCodeExample()
     {
         // Arrange
-        var defaultStressManagementStrategy = (int itemWorryLevel) => itemWorryLevel / 3;
+        var defaultStressManagementStrategy = (long itemWorryLevel) => itemWorryLevel / 3;
         var game = new KeepAwayGame(defaultStressManagementStrategy);
         game.Monkeys.Add(new()
         {
-            Items = new List<int> { 79, 98 },
+            Items = new List<long> { 79, 98 },
             Operation = new Operation("old * 19"),
             ItemTest = new ItemTest(23, 2, 3),
         });
         game.Monkeys.Add(new()
         {
-            Items = new List<int> { 54, 65, 75, 74 },
+            Items = new List<long> { 54, 65, 75, 74 },
             Operation = new Operation("old + 6"),
             ItemTest = new ItemTest(19, 2, 0),
         });
         game.Monkeys.Add(new()
         {
-            Items = new List<int> { 79, 60, 97 },
+            Items = new List<long> { 79, 60, 97 },
             Operation = new Operation("old * old"),
             ItemTest = new ItemTest(13, 1, 3),
         });
         game.Monkeys.Add(new()
         {
-            Items = new List<int> { 74 },
+            Items = new List<long> { 74 },
             Operation = new Operation("old + 3"),
             ItemTest = new ItemTest(17, 0, 1),
         });
